requirements.txt
fasttext
transformers
torch
torchvision
torchaudio
psycopg2-binary
python-dotenv